server/requirements.txt
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.0
dicttoxml==1.7.4
Flask==1.0.2
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
MarkupSafe==1.1.0
Naked==0.1.31
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycryptodome==3.7.2
Pygments==2.3.1
python-decouple==3.1
PyYAML==3.13
requests==2.21.0
shellescape==3.4.1
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
